move puppeteer to a different example
