+++ examples/custom-stream-copy-codec/package.json
