+++ examples/custom-stream-copy-codec/src/config.json
