+++ examples/custom-stream-copy-codec/src/customStream.ts
