+++ examples/custom-stream-copy-codec/src/index.ts
