+++ examples/custom-stream-copy-codec/tsconfig.json
