+++ examples/puppeteer-stream/package.json
@@ -1,15 +1,17 @@
 {
-    "name": "@dank074/discord-video-stream-example",
+    "name": "@dank074/discord-video-stream-example-puppeteer",
     "version": "1.0.0",
     "description": "",
     "main": "dist/index.js",
     "types": "dist/index.d.ts",
     "dependencies": {
         "@dank074/discord-video-stream": "3.0.2",
-        "discord.js-selfbot-v13": "^2.14.9"
+        "discord.js-selfbot-v13": "^2.15.1",
+        "puppeteer": "^19.11.1",
+        "puppeteer-stream": "^2.1.4"
     },
     "devDependencies": {
-        "@types/node": "^18.14.1",
+        "@types/node": "^18.19.8",
         "typescript": "^4.9.5"
     },
     "scripts": {
