+++ examples/puppeteer-stream/src/config.json
@@ -1,0 +1,13 @@
+{
+    "token": "SELF TOKEN HERE",
+    "acceptedAuthors": ["USER_ID_HERE"],
+    "streamOpts": {
+        "width": 1280,
+        "height": 720,
+        "fps": 30,
+        "bitrateKbps": 1000,
+        "maxBitrateKbps": 2500,
+        "hardware_acceleration": false,
+        "videoCodec": "H264"
+    }
+}
